video controls refactor
